Convert to jest test in testAuthenticateOneTimeLogin.js
